Support NumPy 1.24+ (#124)

+++ lacecore/_common/reindexing.py
@@ -8,7 +8,7 @@ def indices_of_original_elements_after_applying_mask(mask):
     produce an array containing the new indices of the original elements. Returns
     -1 as the index of the removed elements.
     """
-    result = np.repeat(np.int(-1), len(mask))
+    result = np.repeat(-1, len(mask))
     result[mask] = np.arange(np.count_nonzero(mask))
     return result
